Moved atomic boolean classes into own files.
